feat: use shared resources to run tests
Resolves NEX-1743.

Shared test resources can save costs and time to run tests.

- Adds shared test resources storage
- Uses a shared PG to run database tests
- Runs tests per file to share resources
