+++ docs/docs/resources/examples/postgresql.yaml
@@ -21,6 +21,3 @@ spec:
 
   maintenanceWindowDow: sunday
   maintenanceWindowTime: 11:00:00
-
-  userConfig:
-    pg_version: "15"
