+++ tests/database_test.go
@@ -1,3 +1,5 @@
+//go:build database
+
 package tests
 
 import (
@@ -17,47 +19,36 @@ func TestDatabase(t *testing.T) {
 	// GIVEN
 	ctx, cancel := testCtx()
 	defer cancel()
-	s := NewSession(ctx, k8sClient)
 
-	pgName := randName("database-pg")
+	pg, release, err := sharedResources.AcquirePostgreSQL(ctx)
+	require.NoError(t, err)
+	defer release()
+
+	// Cleans test afterward
+	s := NewSession(ctx, k8sClient)
+	defer s.Destroy(t)
+
+	pgName := pg.GetName()
 	dbName := randName("database-db")
 	yml, err := loadExampleYaml("database.yaml", map[string]string{
-		"doc[0].metadata.name":  pgName,
-		"doc[0].spec.project":   cfg.Project,
-		"doc[0].spec.cloudName": cfg.PrimaryCloudName,
-
-		"doc[1].metadata.name":    dbName,
-		"doc[1].spec.project":     cfg.Project,
-		"doc[1].spec.serviceName": pgName,
+		"metadata.name":    dbName,
+		"spec.project":     cfg.Project,
+		"spec.serviceName": pgName,
 
 		// Remove 'databaseName' from the initial yaml
-		"doc[1].spec.databaseName": "REMOVE",
+		"spec.databaseName": "REMOVE",
 	})
 	require.NoError(t, err)
-	// Cleans test afterward
-	defer s.Destroy(t)
 
 	// WHEN
 	// Applies given manifest
 	require.NoError(t, s.Apply(yml))
 
 	// Waits kube objects
-	pg := new(v1alpha1.PostgreSQL)
-	require.NoError(t, s.GetRunning(pg, pgName))
-
 	db := new(v1alpha1.Database)
 	require.NoError(t, s.GetRunning(db, dbName))
 
 	// THEN
-	// Validates PostgreSQL
-	pgAvn, err := avnGen.ServiceGet(ctx, cfg.Project, pgName)
-	require.NoError(t, err)
-	assert.Equal(t, pgAvn.ServiceName, pg.GetName())
-	assert.Equal(t, serviceRunningState, pg.Status.State)
-	assert.Contains(t, serviceRunningStatesAiven, pgAvn.State)
-	assert.Equal(t, pgAvn.Plan, pg.Spec.Plan)
-	assert.Equal(t, pgAvn.CloudName, pg.Spec.CloudName)
-
 	// Validates Database
 	dbAvn, err := controllers.GetDatabaseByName(ctx, avnGen, cfg.Project, pgName, dbName)
 	require.NoError(t, err)
@@ -82,35 +73,20 @@ func TestDatabase_databaseName(t *testing.T) {
 	t.Parallel()
 	defer recoverPanic(t)
 
-	// GIVEN
 	ctx, cancel := testCtx()
 	defer cancel()
+
+	// GIVEN
+	pg, release, err := sharedResources.AcquirePostgreSQL(ctx)
+	require.NoError(t, err)
+	defer release()
+
 	s := NewSession(ctx, k8sClient)
 
 	// Cleans test afterward
 	defer s.Destroy(t)
 
-	pgName := randName("databasename-pg")
-
-	// Create a PostgreSQL to reuse with database test cases
-	yml, err := loadExampleYaml("database.yaml", map[string]string{
-		"doc[0].metadata.name":  pgName,
-		"doc[0].spec.project":   cfg.Project,
-		"doc[0].spec.cloudName": cfg.PrimaryCloudName,
-
-		"doc[1]": "REMOVE", // remove database from yaml
-	})
-	require.NoError(t, err)
-
-	// WHEN
-	// Applies given manifest
-	err = s.Apply(yml)
-	require.NoError(t, err, "Failed to apply YAML")
-
-	// Waits kube objects
-	pg := new(v1alpha1.PostgreSQL)
-	require.NoError(t, s.GetRunning(pg, pgName))
-
+	pgName := pg.GetName()
 	testCases := []struct {
 		name                      string
 		metaDBName                string
@@ -157,18 +133,11 @@ func TestDatabase_databaseName(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			ctx, cancel := testCtx()
-			defer cancel()
-
 			yml, err := loadExampleYaml("database.yaml", map[string]string{
-				"doc[0].metadata.name":  pgName,
-				"doc[0].spec.project":   cfg.Project,
-				"doc[0].spec.cloudName": cfg.PrimaryCloudName,
-
-				"doc[1].metadata.name":     tc.metaDBName,
-				"doc[1].spec.project":      cfg.Project,
-				"doc[1].spec.serviceName":  pgName,
-				"doc[1].spec.databaseName": tc.specDatabaseName, // Set databaseName from test case
+				"metadata.name":     tc.metaDBName,
+				"spec.project":      cfg.Project,
+				"spec.serviceName":  pgName,
+				"spec.databaseName": tc.specDatabaseName, // Set databaseName from test case
 			})
 			require.NoError(t, err)
