+++ tests/kafka_test.go
@@ -1,3 +1,5 @@
+//go:build kafka
+
 package tests
 
 import (
