+++ tests/kafkaacl_test.go
@@ -1,3 +1,5 @@
+//go:build kafkaacl
+
 package tests
 
 import (
