+++ tests/kafkschemaregistryacl_test.go
@@ -1,3 +1,5 @@
+//go:build kafkaschemaregistryacl
+
 package tests
 
 import (
