+++ tests/mysql_test.go
@@ -1,3 +1,5 @@
+//go:build mysql
+
 package tests
 
 import (
