+++ tests/redis_test.go
@@ -1,3 +1,5 @@
+//go:build redis
+
 package tests
 
 import (
