+++ tests/service_opts_test.go
@@ -1,3 +1,5 @@
+//go:build misc
+
 package tests
 
 import (
