+++ tests/serviceintegration_test.go
@@ -1,3 +1,5 @@
+//go:build serviceintegration
+
 package tests
 
 import (
