+++ tests/suite_lib.go
@@ -1,3 +1,5 @@
+//go:build suite
+
 package tests
 
 import (
